backend/requirements.txt
asgiref==3.4.1
Django==3.2.9
django-cors-headers==3.11.0
django-extensions==3.1.5
django-filter==22.1
djangorestframework==3.12.4
drf-nested-routers==0.93.4
freezegun==1.1.0
psycopg2==2.9.1
python-dateutil==2.8.2
pytz==2021.3
PyYAML==6.0
six==1.16.0
sqlparse==0.4.2
uritemplate==4.1.1
